# ssa: optimizes slice allocations